requirements.txt
netmiko
ntc-templates
pyyaml
